Add environment switching logic
Code was added because access to
onyx server was disrupted. Therefore,
I needed an easy to switch between
using environmental variables pointing
to the onyx mysql database and my local
machine mysql database.

During creation of the mysql connection,
the program looks for a file `dev.env`
and if it finds it, then the local mysql
database is used for the program run. If
the program does not find the file, it uses
the onyx server mysql database.